few corrections - stable version

+++ include/common/types.h
@@ -12,8 +12,8 @@ typedef unsigned short uint16_t;
 typedef int int32_t;
 typedef unsigned int uint32_t;
 
-typedef long int int64_t;
-typedef unsigned long int uint64_t;
+typedef long long int int64_t;
+typedef unsigned long long int uint64_t;
 
 typedef uint32_t    size_t;
 }
